Changing api calls to function calls, importing api module into shared

+++ api/src/main/kotlin/com/grup/interfaces/IUserBalanceRepository.kt
@@ -1,7 +1,6 @@
 package com.grup.interfaces
 
 import com.grup.models.UserBalance
-import java.math.BigDecimal
 
 internal interface IUserBalanceRepository {
     fun createUserBalance(userBalance: UserBalance): UserBalance?
@@ -9,5 +8,5 @@ internal interface IUserBalanceRepository {
     fun findUserBalanceByUserAndGroupId(userId: String, groupId: String): UserBalance?
     fun findUserBalancesByGroupId(groupId: String): List<UserBalance>
 
-    fun updateUserBalance(newBalance: BigDecimal): UserBalance?
+    fun updateUserBalance(newBalance: Double): UserBalance?
 }
